Garth dev ward admin

+++ src/database/DBConnection.java
@@ -21,7 +21,8 @@ public class DBConnection {
     private Statement statement;
     private ResultSet resultSet;
 
-    public DBConnection() {}
+    public DBConnection() {
+    }
 
     /* Remote AWS database connection */
     private void getDBConnection() {
@@ -31,6 +32,32 @@ public class DBConnection {
 
             statement = connection.createStatement();
 
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+
+    private void closeConnection() {
+        try {
+            connection.close();
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+    }
+
+    private void closeStatement() {
+        try {
+            if (statement != null)
+                statement.close();
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+
+    private void closeResultSet() {
+        try {
+            if (resultSet != null)
+                resultSet.close();
         } catch (Exception e) {
             e.printStackTrace();
         }
@@ -59,8 +86,8 @@ public class DBConnection {
                 Double onHoliday = resultSet.getDouble("onHoliday");
                 Double offSick = resultSet.getDouble("offSick");
                 int ward_ID = resultSet.getInt("ward_ID");
-
                 String password = resultSet.getString("password");
+                String employee_type = resultSet.getString("employee_type");
                 String privilege = resultSet.getString("privilege");
 
                 Employee employee = new Employee();
@@ -87,6 +114,7 @@ public class DBConnection {
                 employee.setOffSick(offSick);
                 employee.setWard_ID(ward_ID);
                 employee.setPassword(password);
+                employee.setEmployee_type(employee_type);
                 employee.setPrivilege(privilege);
 
                 employeeList.add(employee);
@@ -103,13 +131,13 @@ public class DBConnection {
     }
 
     public void createEmployee(String fNameIn, String sNameIn, Calendar DOBIn, String contactNumIn, String emailIn,
-                               double numHolidaysIn, double contractHoursIn, double salary, int ward_IDIn, String password, String privilege) {
+                               double numHolidaysIn, double contractHoursIn, double salary, int ward_IDIn, String password, String privilege, String employee_type) {
 
         getDBConnection();
 
         try {
             PreparedStatement preparedStatement;
-            preparedStatement = connection.prepareStatement("INSERT INTO employee VALUES(?,?,?,?,?,?,?,?,?,?,?,?,?,?,?);");
+            preparedStatement = connection.prepareStatement("INSERT INTO employee VALUES(?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?);");
             preparedStatement.setInt(1, 0); // employee id 0 as it's auto incremented in DB
             preparedStatement.setString(2, fNameIn);
             preparedStatement.setString(3, sNameIn);
@@ -128,7 +156,7 @@ public class DBConnection {
             preparedStatement.setInt(13, ward_IDIn);
             preparedStatement.setString(14, password);
             preparedStatement.setString(15, privilege);
-//            System.out.print("Details; \n" + fNameIn + " " + sNameIn + " " + sqlDate.toString() + " " + contactNumIn + " " + emailIn + " " + numHolidaysIn + " " + contractHoursIn + " " + salary + " " + ward_IDIn);
+            preparedStatement.setString(16, employee_type);
             preparedStatement.executeUpdate();
 
         } catch (Exception e) {
@@ -141,30 +169,65 @@ public class DBConnection {
         }
     }
 
-    public void removeEmployee(Employee employee) {
+    public boolean removeEmployee(Employee employee) {
+        boolean shiftsAssigned = false;
+        shiftsAssigned = removeEmployeeFromShift_Employee(employee);
+        boolean employeeRemoved = false;
+
+        if(!shiftsAssigned) {
+            getDBConnection();
+
+            try {
+                String name = employee.getfName();
+                PreparedStatement preparedStatement;
+                preparedStatement = connection.prepareStatement("DELETE from employee WHERE emp_ID = ?;");
+                preparedStatement.setInt(1, employee.getEmp_ID());
+                int count = preparedStatement.executeUpdate();
+
+                if (count > 0) {
+                    JOptionPane.showMessageDialog(null, name + " has been removed.", "Success", JOptionPane.INFORMATION_MESSAGE);
+                    employeeRemoved = true;
+                }
+
+            } catch (Exception e) {
+                e.getStackTrace();
+            } finally {
+                closeResultSet();
+                closeStatement();
+                closeConnection();
+            }
+        }
+
+        return employeeRemoved;
+
+    }
+
+    private boolean removeEmployeeFromShift_Employee(Employee employee) {
+
+        boolean shiftsAssigned = false;
 
         getDBConnection();
 
         try {
-            String name = employee.getfName();
             PreparedStatement preparedStatement;
-            preparedStatement = connection.prepareStatement("DELETE from employee WHERE emp_ID = ?;");
+            preparedStatement = connection.prepareStatement("SELECT * from shift_employee WHERE emp_ID = ?;");
             preparedStatement.setInt(1, employee.getEmp_ID());
-            int count = preparedStatement.executeUpdate();
+            int results = preparedStatement.executeUpdate();
 
-            if (count > 0)
-                JOptionPane.showMessageDialog(null, name + " has been removed.", "Success", JOptionPane.INFORMATION_MESSAGE);
+            if (results > 0)
+                shiftsAssigned = true;
             else
-                JOptionPane.showMessageDialog(null, "Employee not removed.", "Error", JOptionPane.ERROR_MESSAGE);
+                shiftsAssigned = false;
 
         } catch (Exception e) {
             e.getStackTrace();
-        }
-        finally {
+        } finally {
             closeResultSet();
             closeStatement();
             closeConnection();
         }
+
+        return shiftsAssigned;
     }
 
     public void updateEmployee(Employee employee) {
@@ -175,7 +238,7 @@ public class DBConnection {
             PreparedStatement preparedStatement;
             preparedStatement = connection.prepareStatement("UPDATE employee SET fName = ?, sName = ?, " +
                     "DOB = ?, contactNum = ?, email = ?, numHolidays = ?, contractHours = ?, " +
-                    "salary = ?, onHoliday = ?, offSick = ?, lastShift = ?, ward_ID = ?, password = ?, privilege = ? WHERE emp_ID = " + employee.getEmp_ID() + ";");
+                    "salary = ?, onHoliday = ?, offSick = ?, lastShift = ?, ward_ID = ?, password = ?, privilege = ?, employee_type = ? WHERE emp_ID = " + employee.getEmp_ID() + ";");
             preparedStatement.setString(1, employee.getfName());
             preparedStatement.setString(2, employee.getsName());
 
@@ -193,6 +256,7 @@ public class DBConnection {
             preparedStatement.setInt(12, employee.getWard_ID());
             preparedStatement.setString(13, employee.getPassword());
             preparedStatement.setString(14, employee.getPrivilege());
+            preparedStatement.setString(15, employee.getEmployee_type());
             int count = preparedStatement.executeUpdate();
 
             if (count > 0)
@@ -247,57 +311,37 @@ public class DBConnection {
         return resultList;
     }
 
-    private void closeConnection() {
-        try {
-            connection.close();
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
-    }
-
-    private void closeStatement() {
-        try {
-            if (statement != null)
-                statement.close();
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-    }
-
-    private void closeResultSet() {
-        try {
-            if (resultSet != null)
-                resultSet.close();
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-    }
-
-    public ArrayList<Ward> getWards(){
-        String query="select * from ward;";
-        ArrayList<Ward> wards=new ArrayList<>();
+    public ArrayList<Ward> getWards() {
+        getDBConnection();
+        String query = "select * from ward;";
+        ArrayList<Ward> wards = new ArrayList<>();
         try {
             resultSet = statement.executeQuery(query);
-            while(resultSet.next()){
-                Ward ward=new Ward();
+            while (resultSet.next()) {
+                Ward ward = new Ward();
                 ward.setWard_ID(resultSet.getInt("ward_ID"));
+                ward.setWardType(resultSet.getString("wardType"));
                 ward.setReqNurses(resultSet.getInt("reqNurses"));
                 ward.setReqDoctors(resultSet.getInt("reqDoctors"));
                 wards.add(ward);
             }
-        }
-        catch (Exception e){
+        } catch (Exception e) {
             e.printStackTrace();
+        } finally {
+            closeResultSet();
+            closeStatement();
+            closeConnection();
         }
         return wards;
     }
-    public ArrayList<Shift> getShifts(){
-        String query="select * from shift;";
-        ArrayList<Shift> shifts=new ArrayList<>();
+
+    public ArrayList<Shift> getShifts() {
+        String query = "select * from shift;";
+        ArrayList<Shift> shifts = new ArrayList<>();
         try {
             resultSet = statement.executeQuery(query);
-            while(resultSet.next()){
-                Shift shift=new Shift();
+            while (resultSet.next()) {
+                Shift shift = new Shift();
                 shift.setShift_ID(resultSet.getInt("shift_ID"));
                 shift.setStartTime(resultSet.getString("startTime"));
                 shift.setEndTime(resultSet.getString("endTime"));
@@ -306,10 +350,87 @@ public class DBConnection {
                 shift.setDayOfWeek(resultSet.getString("dayOfWeek"));
                 shifts.add(shift);
             }
-        }
-        catch (Exception e){
+        } catch (Exception e) {
             e.printStackTrace();
+        } finally {
+            closeResultSet();
+            closeStatement();
+            closeConnection();
         }
         return shifts;
+    }
+
+    public void addWard(Ward ward) {
+
+        getDBConnection();
+
+        try {
+            PreparedStatement preparedStatement;
+            preparedStatement = connection.prepareStatement("INSERT INTO ward VALUES(?,?,?,?);");
+            preparedStatement.setInt(1, 0); // employee id 0 as it's auto incremented in DB
+            preparedStatement.setString(2, ward.getWardType());
+            preparedStatement.setInt(3, ward.getReqNurses());
+            preparedStatement.setInt(4, ward.getReqDoctors());
+            preparedStatement.executeUpdate();
+
+        } catch (Exception e) {
+            e.printStackTrace();
+        } finally {
+            closeResultSet();
+            closeStatement();
+            closeConnection();
+        }
+    }
+
+    public void removeWard(Ward ward) {
+
+        getDBConnection();
+
+        try {
+            String wardType = ward.getWardType();
+            PreparedStatement preparedStatement;
+            preparedStatement = connection.prepareStatement("DELETE from ward WHERE ward_ID = ?;");
+            preparedStatement.setInt(1, ward.getWard_ID());
+            int count = preparedStatement.executeUpdate();
+
+            if (count > 0)
+                JOptionPane.showMessageDialog(null, wardType + " has been removed.", "Success", JOptionPane.INFORMATION_MESSAGE);
+            else
+                JOptionPane.showMessageDialog(null, "Ward not removed.", "Error", JOptionPane.ERROR_MESSAGE);
+
+        } catch (Exception e) {
+            e.getStackTrace();
+        } finally {
+            closeResultSet();
+            closeStatement();
+            closeConnection();
+        }
+    }
+
+    public void updateWard(Ward ward) {
+
+        getDBConnection();
+
+        try {
+            PreparedStatement preparedStatement;
+            preparedStatement = connection.prepareStatement("UPDATE ward SET wardType = ?, reqNurses = ?, " +
+                    "reqDoctors = ? WHERE ward_ID = " + ward.getWard_ID() + ";");
+            preparedStatement.setString(1, ward.getWardType());
+            preparedStatement.setInt(2, ward.getReqNurses());
+            preparedStatement.setInt(3, ward.getReqDoctors());
+            int count = preparedStatement.executeUpdate();
+
+            if (count > 0)
+                JOptionPane.showMessageDialog(null, "Ward updated.", "Success", JOptionPane.INFORMATION_MESSAGE);
+            else
+                JOptionPane.showMessageDialog(null, "Ward not updated.", "Error", JOptionPane.ERROR_MESSAGE);
+
+        } catch (Exception e) {
+            e.printStackTrace();
+        } finally {
+            closeResultSet();
+            closeStatement();
+            closeConnection();
+        }
     }
 }
